Update build artifacts for ProcessorEmulator

+++ IChipsetEmulator.cs
@@ -4,8 +4,8 @@ namespace ProcessorEmulator.Tools
     {
         string ChipsetName { get; }
         bool Initialize(string configPath);
-        byte[] ReadRegister(uint address);
-        void WriteRegister(uint address, byte[] data);
+    byte[] ReadRegister(long address);
+    void WriteRegister(long address, byte[] data);
         // Add other chipset-specific functions here
     }
 }
